fixed serializers and model mixed ups
