+++ windscreen_app/models.py
@@ -1,12 +1,7 @@
 from django.db import models
 from django.utils import timezone  
-from rest_framework.parsers import MultiPartParser, FormParser
-from rest_framework.response import Response
-from rest_framework.decorators import action
-from rest_framework import viewsets, status
-from rest_framework.views import APIView
 from django.core.validators import FileExtensionValidator
-from windscreen_app.serializers import WorkProgressSerializer
+
 
 class Vehicle(models.Model):
     registration_number = models.CharField(max_length=15, unique=True)
@@ -33,6 +28,7 @@ class Quote(models.Model):
     services = models.ManyToManyField('Service')
     total_cost = models.DecimalField(max_digits=10, decimal_places=2)
     status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='Pending')
+    registration_number = models.CharField(max_length=50, blank=True, null=True)
 
     def __str__(self):
         return self.quote_number
@@ -114,28 +110,7 @@ class WorkProgress(models.Model):
     def __str__(self):
         return f"{self.vehicle_reg_no} - Order: {self.order.order_number} - Services: {', '.join(service.name for service in self.services.all())}"
 
-class WorkProgressViewSet(viewsets.ModelViewSet):
-    queryset = WorkProgress.objects.all().order_by('-created_at')
-    serializer_class = WorkProgressSerializer
-    parser_classes = (MultiPartParser, FormParser)  
-
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    # Optional: Endpoint to filter by vehicle registration number
-    @action(detail=False, methods=['get'], url_path='filter')
-    def filter_by_vehicle(self, request):
-        reg_no = request.query_params.get('vehicle_reg_no')
-        if reg_no:
-            queryset = self.get_queryset().filter(vehicle_reg_no=reg_no)
-            serializer = self.get_serializer(queryset, many=True)
-            return Response(serializer.data)
-        return Response({"error": "Please provide a vehicle_reg_no parameter."}, 
-                        status=status.HTTP_400_BAD_REQUEST)    
+   
 
 
 #invoice model
