+++ windscreen_app/views.py
@@ -152,7 +152,7 @@ class SubmitServiceAPIView(APIView):
 
 class GetQuotesAPIView(APIView):
     def get(self, request):
-        quotes = Quote.objects.exclude(status="Rejected")  # Exclude rejected quotes
+        quotes = Quote.objects.all() # Exclude rejected quotes
         serializer = QuoteSerializer(quotes, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
@@ -264,3 +264,27 @@ class InvoiceViewSet(viewsets.ModelViewSet):
             queryset = queryset.filter(status=status)
 
         return queryset
+
+
+class WorkProgressViewSet(viewsets.ModelViewSet):
+    queryset = WorkProgress.objects.all().order_by('-created_at')
+    serializer_class = WorkProgressSerializer
+    parser_classes = (MultiPartParser, FormParser)  
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    # Optional: Endpoint to filter by vehicle registration number
+    @action(detail=False, methods=['get'], url_path='filter')
+    def filter_by_vehicle(self, request):
+        reg_no = request.query_params.get('vehicle_reg_no')
+        if reg_no:
+            queryset = self.get_queryset().filter(vehicle_reg_no=reg_no)
+            serializer = self.get_serializer(queryset, many=True)
+            return Response(serializer.data)
+        return Response({"error": "Please provide a vehicle_reg_no parameter."}, 
+                        status=status.HTTP_400_BAD_REQUEST)
